requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.1
flask-restplus==0.13.0
flask==1.1.2importlib-metadata==1.6.0; python_version < '3.8'
itsdangerous==1.1.0
jinja2==2.11.1
jsonschema==3.2.0
markupsafe==1.1.1
pyrsistent==0.16.0
python-simplexml==0.1.5
pytz==2019.3
six==1.14.0
werkzeug==1.0.1
zipp==3.1.0